Extracted the adapter
Issue #6
