+++ .idea/fileTemplates/includes/File Header.java
@@ -1,0 +1,6 @@
+/**
+ * Created by ${USER} on ${DAY}.${MONTH}.${YEAR}.
+ *
+ * Copyright (c) 2017 Gini GmbH.
+ */
+
